+++ app/build.gradle.kts
@@ -68,7 +68,9 @@ dependencies {
     debugImplementation(libs.androidx.ui.tooling)
     implementation("androidx.compose.ui:ui:1.5.3")
     //implementation("androidx.navigation:navigation-compose:2.6.0")
-    implementation("androidx.compose.material3:material3:1.1.1")
+    implementation ("androidx.compose.material3:material3:1.3.0")
+    implementation ("androidx.compose.material3:material3-window-size-class:1.3.0")
+
     implementation("androidx.compose.ui:ui-tooling-preview:1.5.3")
 
     implementation("androidx.compose.material:material:1.4.0") // Reemplaza con la versión correspondiente
