+++ app/src/main/java/com/ita/condominio/Components.kt
@@ -1,0 +1,90 @@
+package com.ita.condominio
+
+import androidx.compose.foundation.layout.*
+import androidx.compose.foundation.shape.RoundedCornerShape
+import androidx.compose.material.BottomNavigation
+import androidx.compose.material.BottomNavigationItem
+import androidx.compose.material3.*
+import androidx.compose.material3.Icon
+import androidx.compose.material3.Text
+import androidx.compose.runtime.Composable
+import androidx.compose.ui.Modifier
+import androidx.compose.ui.graphics.Color
+import androidx.compose.ui.res.painterResource
+import androidx.compose.ui.unit.dp
+import androidx.navigation.NavHostController
+
+@Composable
+fun AccountOptionButton(text: String, iconRes: Int, onClick: () -> Unit) {
+    Button(
+        onClick = onClick,
+        modifier = Modifier
+            .fillMaxWidth()
+            .height(60.dp)
+            .padding(horizontal = 20.dp, vertical = 8.dp),
+        shape = RoundedCornerShape(10.dp),
+        colors = ButtonDefaults.buttonColors(containerColor = Color(0xFFa9dfbf))
+    ) {
+        Icon(
+            painter = painterResource(id = iconRes),
+            contentDescription = null,
+            modifier = Modifier.size(25.dp),
+            tint = Color.Black
+        )
+        Spacer(modifier = Modifier.width(8.dp))
+        Text(text = text, color = Color.Black)
+    }
+}
+
+@Composable
+fun BottomNavigationBar(navController: NavHostController) {
+    BottomNavigation(
+        backgroundColor = Color(0xFFC4D9D2),
+        contentColor = Color.Black
+    ) {
+        BottomNavigationItem(
+            icon = {
+                Icon(
+                    painter = painterResource(id = R.drawable.reporte),
+                    contentDescription = "Informes",
+                    modifier = Modifier.size(30.dp)
+                )
+            },
+            selected = false,
+            onClick = { /* Navegar a informes */ }
+        )
+        BottomNavigationItem(
+            icon = {
+                Icon(
+                    painter = painterResource(id = R.drawable.home),
+                    contentDescription = "Home",
+                    modifier = Modifier.size(30.dp)
+                )
+            },
+            selected = false,
+            onClick = { /* Navegar a Home */ }
+        )
+        BottomNavigationItem(
+            icon = {
+                Icon(
+                    painter = painterResource(id = R.drawable.chat),
+                    contentDescription = "Chat",
+                    modifier = Modifier.size(30.dp)
+                )
+            },
+            selected = false,
+            onClick = { /* Navegar a Chat */ }
+        )
+        BottomNavigationItem(
+            icon = {
+                Icon(
+                    painter = painterResource(id = R.drawable.user),
+                    contentDescription = "Usuario",
+                    modifier = Modifier.size(30.dp)
+                )
+            },
+            selected = false,
+            onClick = { /* Navegar a Usuario */ }
+        )
+    }
+}
