--- app/src/main/java/com/ita/condominio/components.kt
@@ -1,2 +1,0 @@
-package com.ita.condominio
-
